requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.7b0
boto3==1.17.87
botocore==1.20.87
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-appconf==1.0.4
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.0.0
django-imagekit==4.0.2
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.0
pandas==1.3.2
pathspec==0.8.1
phonenumbers==8.12.23
pilkit==2.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.0
urllib3==1.26.4
whitenoise==5.2.0
